+++ src/pipecat/processors/idle_frame_processor.py
@@ -11,6 +11,7 @@ from typing import Awaitable, Callable, List, Optional
 
 from pipecat.frames.frames import Frame, StartFrame
 from pipecat.processors.frame_processor import FrameDirection, FrameProcessor
+from pipecat.utils.asyncio.watchdog_event import WatchdogEvent
 
 
 class IdleFrameProcessor(FrameProcessor):
@@ -77,7 +78,7 @@ class IdleFrameProcessor(FrameProcessor):
     def _create_idle_task(self):
         """Create and start the idle monitoring task."""
         if not self._idle_task:
-            self._idle_event = asyncio.Event()
+            self._idle_event = WatchdogEvent(self.task_manager)
             self._idle_task = self.create_task(self._idle_task_handler())
 
     async def _idle_task_handler(self):