+++ src/pipecat/services/tavus/video.py
@@ -17,7 +17,7 @@ import aiohttp
 from daily.daily import AudioData, VideoFrame
 from loguru import logger
 
-from pipecat.audio.utils import create_default_resampler
+from pipecat.audio.utils import create_stream_resampler
 from pipecat.frames.frames import (
     CancelFrame,
     EndFrame,
@@ -42,6 +42,7 @@ class TavusVideoService(AIService):
     are routed through Pipecat's media pipeline.
 
     In use cases with DailyTransport, this creates two distinct virtual rooms:
+
     - Tavus room: Contains the Tavus Avatar and the Pipecat Bot
     - User room: Contains the Pipecat Bot and the user
     """
@@ -74,7 +75,7 @@ class TavusVideoService(AIService):
         self._client: Optional[TavusTransportClient] = None
 
         self._conversation_id: str
-        self._resampler = create_default_resampler()
+        self._resampler = create_stream_resampler()
 
         self._audio_buffer = bytearray()
         self._send_task: Optional[asyncio.Task] = None
@@ -243,6 +244,7 @@ class TavusVideoService(AIService):
     async def _cancel_send_task(self):
         """Cancel the audio sending task if it exists."""
         if self._send_task:
+            self._queue.cancel()
             await self.cancel_task(self._send_task)
             self._send_task = None